Continued restyling for consistency
Redid design for HomePage, began styling for forms, divided Burial Registration into multiple components
