+++ package.json
@@ -12,6 +12,7 @@
     "@testing-library/user-event": "^14.4.3",
     "axios": "^1.4.0",
     "bootstrap": "^5.2.3",
+    "bootstrap-icons": "^1.10.5",
     "convert-csv-to-array": "^1.0.3",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
@@ -21,10 +22,12 @@
     "react-router-dom": "^6.8.1",
     "react-scripts": "^5.0.1",
     "react-search-box": "^3.0.0",
+    "react-slick": "^0.29.0",
     "react-stripe-checkout": "^2.6.3",
     "react-toastify": "^9.1.2",
     "react-tooltip": "^5.11.1",
     "seamless-scroll-polyfill": "^2.3.4",
+    "slick-carousel": "^1.8.1",
     "web-vitals": "^3.1.1"
   },
   "scripts": {
@@ -53,5 +56,10 @@
   },
   "devDependencies": {
     "tailwindcss": "^3.3.3"
+  },
+  "prettier": {
+    "singleQuote": true,
+    "tabWidth": 2
   }
+  
 }
