+++ src/App.css
@@ -4,66 +4,5 @@
 }
 
 body {
-  background-color: #faf8e3;
+  background-color: #fffef1;
 }
-/*
-.FormGroup {
-	margin: 0 15px 20px;
-	padding: 0;
-	border-style: none;
-	background-color: #7795f8;
-	will-change: opacity, transform;
-	box-shadow: 0 6px 9px rgba(50, 50, 93, 0.06), 0 2px 5px rgba(0, 0, 0, 0.08), inset 0 1px 0 #829fff;
-	border-radius: 4px;
-}
-
-.FormRow {
-	display: -ms-flexbox;
-	display: flex;
-	-ms-flex-align: center;
-	align-items: center;
-	margin-left: 15px;
-	border-top: 1px solid #819efc;
-}
-
-.StripeElement--webkit-autofill {
-	background: transparent !important;
-}
-
-.StripeElement {
-	width: 100%;
-	padding: 11px 15px 11px 0;
-}
-
-button {
-	display: block;
-	font-size: 16px;
-	width: calc(100% - 30px);
-	height: 40px;
-	margin: 40px 15px 0;
-	background-color: hsla(130, 89%, 78%, 0.809);
-	box-shadow: 0 6px 9px rgba(50, 50, 93, 0.06), 0 2px 5px rgba(0, 0, 0, 0.08), inset 0 1px 0 #ffb9f6;
-	border-radius: 4px;
-	color: rgba(255, 255, 255, 0.981);
-	font-weight: 600;
-	cursor: pointer;
-	transition: all 100ms ease-in-out;
-	will-change: transform, background-color, box-shadow;
-	border: none;
-}
-
-button:active {
-	background-color: #83d982;
-	box-shadow: 0 6px 9px rgba(50, 50, 93, 0.06), 0 2px 5px rgba(0, 0, 0, 0.08), inset 0 1px 0 #e298d8;
-	transform: scale(0.99);
-}
-
-/*
-.pages {
-  color: #ace1ff;
-  text-align: center;
-  font-size: 3.5rem;
-  margin-top: 10%;
-}
-
-*/
