+++ src/Pages/BurialRegistration/PaymentForm.js
@@ -1,0 +1,67 @@
+import axios from 'axios';
+import React from 'react';
+// import StripeCheckout from 'react-stripe-checkout';
+import { useState } from 'react';
+// import { toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+import './BurialReg.css';
+import { CardElement, useElements, useStripe } from '@stripe/react-stripe-js';
+
+// import { ToastContainer } from 'react-toastify';
+// import { otherPropsFromToastConfigure } from 'react-toastify';
+// import StripeContainer from "./StripeContainer";
+
+export default function PaymentForm() {
+  const [success, setSuccess] = useState(false);
+  const stripe = useStripe();
+  const elements = useElements();
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    const { error, paymentMethod } = await stripe.createPaymentMethod({
+      type: 'card',
+      card: elements.getElement(CardElement),
+    });
+
+    if (!error) {
+      try {
+        const { id } = paymentMethod;
+        const response = await axios.post(
+          'http://localhost:3000/zeForm/Payment',
+          {
+            amount: 1000,
+            id,
+          }
+        );
+
+        if (response.data.success) {
+          console.log('Successful payment');
+          setSuccess(true);
+        }
+      } catch (error) {
+        console.log('Error', error);
+      }
+    } else {
+      console.log(error.message);
+    }
+  };
+
+  return (
+    <div className="p-12 lg:px-64 pb-24">
+      {!success ? (
+        <form onSubmit={handleSubmit}>
+          <fieldset className="FormGroup">
+            <div className="FormRow">
+              <CardElement />
+            </div>
+          </fieldset>
+          <button>Pay</button>
+        </form>
+      ) : (
+        <div>
+          <h2>Payment Successful! See you at the graveyard. </h2>
+        </div>
+      )}
+    </div>
+  );
+}
