--- src/Pages/Home/Home.css
@@ -1,89 +1,0 @@
-/* WARNING: Any changes made to this file will be lost when your Formidable settings are updated. */
-#header{
-    display: flex;
-    justify-content: center;
-    vertical-align: middle;
-    /* max-width: ; */
-    /* align-content: center; */
-    /* flex-direction: column; */
-}
-#overlay{
-    background: black;
-    overflow: hidden;
-}
-
-.about-us{
-    height: max-content;
-    width: 100%;
-    padding: 90px 0;
-    background: #ddd;
-  }
-  .pic{
-    height: auto;
-    width:  302px;
-  }
-  #gate-img{
-    opacity: 0.6;
-  }
-  .about{
-    width: 1130px;
-    max-width: 85%;
-    margin: 0 auto;
-    display: flex;
-    align-items: center;
-    justify-content: space-around;
-  }
-  .text{
-    width: 45vw;
-  }
-  .text h2{
-    font-size: 90px;
-    font-weight: 600;
-    margin-bottom: 10px;
-  
-  }
-  .text h5{
-    font-size: 22px;
-    font-weight: 500;
-    margin-bottom: 20px;
-  }
-  span{
-    color: #4070f4;
-  }
-  .text p{
-    font-size: 18px;
-    line-height: 25px;
-    letter-spacing: 1px;
-  }
-  .data{
-    margin-top: 30px;
-  }
-  .hire{
-    font-size: 18px;
-    background: #4070f4;
-    color: #fff;
-    text-decoration: none;
-    border: none;
-    padding: 8px 25px;
-    border-radius: 6px;
-    transition: 0.5s;
-  }
-  .hire:hover{
-    background: #000;
-    border: 1px solid #4070f4;
-  }
-
-main{
-    background-color: #FAF9F6    ;
-}
-#car-img{
-    width: 100% !important;
-    height: 70vh;
-}
-#buttons{
-    display: flex;
-    justify-content: center;
-}
-#aya{
-    text-align: center;
-}
