+++ src/components/Footer/Footer.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 const Footer = () => {
   return (
@@ -7,14 +7,28 @@ const Footer = () => {
         <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-auto">
           <div className="text-left mb-4 pl-8 sm:mb-0 sm:px-4 ">
             <h2 className="text-white text-xl title-font mb-3">Quick Links</h2>
-            <a href="/" className="text-white text-font hover:text-green-400 block mb-2">Home</a>
-            <a href="/contact" className="text-white text-font hover:text-green-400 block mb-2">Contact Us</a>
+            <a
+              href="/"
+              className="text-white text-font hover:text-green-400 block mb-2"
+            >
+              Home
+            </a>
+            <a
+              href="/contact"
+              className="text-white text-font hover:text-green-400 block mb-2"
+            >
+              Contact Us
+            </a>
             <p className="text-white text-font">123 Street, City, Country</p>
           </div>
           <div className="text-left mb-4 pl-8 sm:mb-0 sm:px-4">
             <h2 className="text-white text-xl title-font mb-2">Contact Info</h2>
-            <p className="text-white text-font hover:text-green-400 block mb-2">email@email.org</p>
-            <p className="text-white text-font hover:text-green-400 block mb-2">(111) 111-1111</p>
+            <p className="text-white text-font hover:text-green-400 block mb-2">
+              email@email.org
+            </p>
+            <p className="text-white text-font hover:text-green-400 block mb-2">
+              (111) 111-1111
+            </p>
           </div>
           <div className="text-left sm:px-4 sm:pl-8 mt-auto">
             <ul className="flex space-x-4">
@@ -42,7 +56,8 @@ const Footer = () => {
           </div>
         </div>
         <div className="text-center mt-4 text-yellow-300">
-          &copy; {new Date().getFullYear()} Five Pillars Cemetery. Created by Joe.
+          &copy; {new Date().getFullYear()} Five Pillars Cemetery. Created by
+          Joe.
         </div>
       </div>
     </div>
