+++ src/components/NavBar/NavBar.css
@@ -1,27 +1,14 @@
-.navbar {
-
-  height: 20%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 1.2rem;
-  position: sticky;
-  top: 0;
-  z-index: 20;
-  width: 100%;
+/* @media screen and (min-width: 960px) { */
+.navbarTitle {
+  text-align: center;
+  color: #faf8e3;
+  font-size: x-large;
+  padding: 10px;
 }
 
-/* @media screen and (min-width: 960px) { */
-  .navbarTitle {
-    text-align: center;
-    color: #faf8e3;
-    font-size: x-large;
-    padding: 10px;
-  }
-
-  .navbarLogoTitle {
-    height: 0;
-  }
+.navbarLogoTitle {
+  height: 0;
+}
 /* } */
 
 @media screen and (min-width: 960px) {
@@ -76,7 +63,7 @@
 }
 
 .nav-item:after {
-  content: "";
+  content: '';
   display: block;
   height: 3px;
   width: 0;
@@ -108,7 +95,6 @@
     left: -110%;
     opacity: 1;
     transition: all 0.5s ease;
-
   }
 
   .nav-menu.active {
@@ -126,7 +112,7 @@
     position: absolute;
     top: 0;
     right: 0;
-    transform: translate(-100%, 60%);
+    transform: translate(-100%, 35%);
     font-size: 1.8rem;
     cursor: pointer;
     color: #ffe365;
